--- src/app/components/signup/signup.component.scss
@@ -1,233 +1,0 @@
-.row{
-    width: 100%;
-}
-
-.topmargin{
-    margin-top: 35px;
-}
-
-.top{
-    color: #8e8e8e;
-    font-weight: 600px;
-    margin-top: 20px;
-    font-size: 17px;
-    text-align: center;
-    margin-bottom: 20px;
-}
-
-.display{
-    display: flex;
-    justify-content: center;
-}
-
-.box1{
-    background-color: white;
-    padding: 30px;
-    margin: 25px;
-    border:1px solid rgb(219,219,219);
-}
-
-.instalogo{
-    width: 180px;
-    height: 50px;
-    display: block;
-    margin-left: auto;
-    margin-right: auto;
-}
-
-input{
-    width: 100%;
-    font-size: 11px;
-    height: 40px;
-    background-color: #fafafa;
-}
-
-input:focus{
-    box-shadow: none;
-    border: 1px solid #817e7e;
-    background-color: #FAFAFA;
-}
-
-.btn{
-    width: 100%;
-    background-color: #0095f6;
-    color: white;
-    font-weight: bold;
-    font-size: 14px;
-}
-
-.floating-label-group {
-	position: relative;
-	margin-top: 2px;
-	margin-bottom: 8px;
-
-	.floating-label {
-		font-size: 13px;
-		color: #cccccc;
-		position: absolute;
-		pointer-events: none;
-		top: 9px;
-		left: 15px;
-		transition: all 0.1s ease;
-        margin-bottom: 3px;
-	}
-
-	input:focus ~ .floating-label,
-	input:not(:focus):valid ~ .floating-label {
-		top: -4px;
-		bottom: 0px;
-		left: 10px;
-		font-size: 10px;
-		opacity: 1;
-		color: grey;
-        background-color: #FAFAFA;
-        margin-bottom: 3px;
-        overflow: hidden;
-        height: 20px;
-	}
-}
-
-.field-icon {
-    float: right;
-    margin-left: -30px;
-    margin-top: -30px;
-    margin-right: 13px;
-    position: relative;
-    z-index: 2;
-    font-size: 14px;
-    font-weight: bold;
-    border-radius: 100%;
-    height: 23px;
-    width: 23px;
-}
-.field-icon i{
-    margin-left: 6px;
-    margin-top: 1px;
-}
-
-.field-icon .fa-check{
-    margin-left: 4px;
-    margin-top: 1px;
-}
-
-.field-icon-password {
-    float: right;
-    margin-left: -15px;
-    margin-top: -30px;
-    margin-right: 45px;
-    position: relative;
-    z-index: 2;
-    font-size: 14px;
-    font-weight: bold;
-    border-radius: 100%;
-    height: 23px;
-    width: 23px;
-}
-
-.field-icon-password .fa-check{
-    margin-left: 4px;
-    margin-top: 1px;
-}
-
-
-.orClass{
-    position: relative;
-    font-size: 13px;
-    margin-top: -28px;
-    padding: 10px;
-}
-
-.orClass p{
-    background-color: white;
-    width: 50px;
-    border-radius: 50px;
-    display: block;
-    margin-left: auto;
-    margin-right: auto;
-    text-align: center;
-    margin-top: -10px;
-    color: grey;
-}
-
-.line{
-    margin-top: 35px;
-    background-color: white;
-}
-
-.facebook{
-    color: white;
-    border-radius: 5px;
-    text-align: center;
-    height: 35px;
-    margin-left: auto;
-    margin-right: auto;
-    background-color: #0095f6;
-    margin-top: 0px;
-}
-
-.facebook .fab{
-    font-size: 20px;
-    margin-top: 7px;
-    margin-right: 8px;
-}
-
-.facebook p{
-    font-size: 15px;
-    font-weight: bold;
-    margin-top: 6px;
-}
-
-.forgot-password{
-    text-align: center;
-    font-size: 12px;
-}
-
-.signup{
-    margin-top: 10px;
-    background-color: white;
-    padding: 13px;
-    margin: 25px;
-    border: 1px solid rgb(219,219,219);
-}
-
-.signup p{
-    margin-top: 7px;
-    font-size: 14px;
-}
-
-.signup a{
-    text-decoration: none;
-    color: #0095f6;
-    margin-top: 7px;
-    margin-left: 2px;
-    font-size: 14px;
-}
-
-.getapp{
-    font-size: 14px;
-    text-align: center;
-    margin-top: 20px;
-    margin-bottom: 0px;
-}
-
-.app{
-    margin-bottom: 50px;
-}
-
-.app img{
-    margin-top: 10px;
-    margin-left: 5px;
-    margin-right: 5px;
-}
-
-.agree{
-    font-size: 12px;
-    line-height: 16px;
-    margin: 20px 40px;
-    color: #8e8e8e;
-    text-align: center;
-}
-
-.agree span{
-    font-weight: bold;
-}
