+++ src/app/main/footer/footer.component.scss
