+++ src/app/main/footer/footer.component.spec.ts
