+++ src/app/main/footer/footer.component.ts
